course final

+++ frontend/admin-dashboard/src/pages/course/CourseDetails.jsx
@@ -17,12 +17,13 @@ import {
 } from "@material-tailwind/react";
 import { AddLessons } from "./AddLessons";
 import { message } from "antd";
+import Swal from 'sweetalert2';
+
 
 export const CourseDetails = () => {
   const { id } = useParams();
 
   const [openAccordion, setOpenAccordion] = useState({});
-  const [showAddLessonForm, setShowAddLessonForm] = useState(false);
   const [lessonTitle, setLessonTitle] = useState("");
 
   const [newOpen, setNewOpen] = useState(false);
@@ -58,17 +59,29 @@ export const CourseDetails = () => {
     fetchHandler();
   }, [id, instructorId ,tableLoading]);
 
-  const handleDelete = async () => {
-    try {
-      await axios.delete(`http://localhost:4200/course/${id}`);
-      console.log("Course deleted successfully!");
-      message.success("Course deleted successfully!");
-    } catch (error) {
-      console.error("Error deleting course:", error);
-      message.error("Error deleting course");
-    }
-  };
+  const navigate = useNavigate();
 
+  const handleDelete = async () => {
+    Swal.fire({
+      text: 'Do you want to delete course?',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Yes, delete it!'
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          await axios.delete(`http://localhost:4200/course/${id}`);
+          console.log("Course deleted successfully!");
+          message.success("Course deleted successfully!");
+          navigate("/my-courses");
+        } catch (error) {
+          console.error("Error deleting course:", error);
+          message.error("Error deleting course");
+        }
+      }
+    });
+  };
   
   const handleLessonSubmit = async (e) => {
     e.preventDefault();
@@ -220,38 +233,6 @@ export const CourseDetails = () => {
               Add Lessons
             </Button>
             )}
-
-                  
-                  {showAddLessonForm && (
-                    <form onSubmit={handleLessonSubmit}>
-                      <div className="w-1/4">
-                        <Typography
-                          variant="small"
-                          color="blue-gray"
-                          className="mb-2 font-semibold"
-                        >
-                          Title
-                        </Typography>
-                        <Input
-                          type="text"
-                          name="title"
-                          value={lessonTitle}
-                          onChange={(e) => setLessonTitle(e.target.value)}
-                        />
-                      </div>
-                     <div className="flex gap-5 w-1/4 justify-end">
-                     <button type="submit" className="hidden md:flex w-fit gap-1 rounded-md items-center p-1 px-6 my-10 font-inter font-medium bg-[#9165A0] border-[#9165A0] hover:bg-white text-white hover:text-black border-[1px] hover:border-black text-[14px] transition-colors duration-500">
-                        Add
-                      </button>
-                      <button
-                        onClick={() => setShowAddLessonForm(false)}
-                        className="hidden md:flex w-fit gap-1 rounded-md items-center p-1 px-3 my-10 font-inter font-medium bg-[#9165A0] border-[#9165A0] hover:bg-white text-white hover:text-black border-[1px] hover:border-black text-[14px] transition-colors duration-500"
-                      >
-                        Cancel
-                      </button>
-                     </div>
-                    </form>
-                  )}
                   <ul>
                     {courseDetails.lessons &&
                       courseDetails.lessons.map((lesson, index) => (
